Moved the auth code into a blueprint. Made an app level base.html template which can be used for all blueprints and created an app level static to holder the usgs header/footer less file. Created a style asset for the auth login page.
