+++ pubs_ui/__init__.py
@@ -45,11 +45,11 @@ bower = Bower(app)
 
 import assets
 
-from pubswh.views import pubswh
+from auth.views import auth
 from manager.views import manager
+from pubswh.views import pubswh
 
+app.register_blueprint(auth)
 app.register_blueprint(pubswh)
 app.register_blueprint(manager,
                        url_prefix='/manager')
-
-import views
