+++ pubs_ui/auth/static/less/auth.less
@@ -1,4 +1,3 @@
 @import "../../../bower_components/bootstrap/less/bootstrap.less";
 
 @import "../../../static/less/usgs_header_footer.less";
-@import "search.less";
