--- pubs_ui/manager/static/less/usgs_header_footer.less
@@ -1,79 +1,0 @@
-
-/* USGS header styles */
-header {
-    display: block !important;
-    height: 72px;
-    width: 100%;
-    background: #000;
-  }
-  #usgs_logo {
-    float: left;
-  }
-  #banner {
-    float: left;
-    height: 72px;
-  }
-  #header_nav {
-    margin: 5px 8px 0 0;
-    float: right;
-  }
-  #header_nav ul {
-    font-family: Verdana;
-    list-style: none;
-	margin:0;
-  }
-
-  #header_nav ul li{
-  margin-bottom:2px;}
-
-  #header_nav ul li a {
-    color: #fff;
-    text-decoration: none;
-    font-size: .9em;
-  }
-  #header_nav ul li a:hover {
-    text-decoration: underline;
-  }
-/* ==========================================================================
-   footer styles
-   ========================================================================== */
-footer {
-  clear: both;
-  position: relative;
-}
-
-footer p{margin:0;}
-
-#usgsfooterbar {
-  width: 100%;
-  height: 20px;
-  background: #666666;
-}
-#usgsfooterbar a {
-  color: #fff;
-  font-family: Verdana;
-  text-decoration: none;
-  font-size: .7em;
-  padding-left: 10px;
-}
-#usgsfooterbar a:hover,
-#usgsfootertext a:hover {
-  text-decoration: underline;
-}
-#usgsfootertext {
-  font-size: .85em;
-  font-family: Verdana;
-  padding: 5px;
-  margin:0;
-}
-#usgsfootertext a {
-  color: #003366;
-  text-decoration: none;
-}
-#usgsfootertext img {
-  display: none;
-}
-#take_pride,
-#usa_gov {
-  position: absolute;
-}
