--- pubs_ui/pubswh/templates/pubswh/login.html
@@ -1,38 +1,0 @@
-{% extends 'pubswh/base.html' %}
-{% block head %}
-	<title>Log in to the Publications Warehouse</title>
-	<meta name="description" content="Log in to the Pubs Warehouse">
-{% endblock head %}
-
-{% block page_style %}
-	<link rel="stylesheet" href="{{ url_for('pubswh.static', filename = 'css/pagination.css') }}">
-{% endblock page_style %}
-
-{% block main %}
-
-{% with messages = get_flashed_messages() %}
-  {% if messages %}
-    <ul class=flashes>
-    {% for message in messages %}
-      <li>{{ message }}</li>
-    {% endfor %}
-    </ul>
-  {% endif %}
-{% endwith %}
-  <h1>Sign In</h1>
-    <p>To access certain Publications Warehouse content and features, you need to log in using your Active Directory Username and Password.</p>
-  {% if error %}
-    <p class=error><strong>Error:</strong> {{ error }}
-  {% endif %}
-  <form style="display: inline;" action="" method="post" name="login">
-      {{ form.hidden_tag() }}
-      {{ form.username.label }} {{ form.username }}
-      {{ form.password.label }} {{ form.password }}
-      <button type="submit" class="btn">Log In</button>
-  </form>
-    <br>
-    <br>
-
-{% endblock %}
-
-{% block url %}<a href="{{ url_for('login_page') }}">http://pubs.er.usgs.gov{{ url_for('login_page') }}</a>{% endblock %}
