+++ pubs_ui/templates/base.html
@@ -3,13 +3,15 @@
 	<head lang="en">
 		<meta charset="UTF-8">
 		<meta name="viewport" content="width=device-width, initial-scale=1">
+
+		{% block meta_data %}{% endblock %}
+
 		<title>{% block title %}USGS Publications Warehouse{% endblock %}</title>
-		{% block usgs_style %}
-			{%  assets "manager_style" %}
+		{% block page_style %}
+			{% assets "usgs_style" %}
 				<link rel="stylesheet" href="{{ ASSET_URL }}" type="text/css" />
 			{% endassets %}
 		{% endblock %}
-		{% block page_style %}{% endblock %}
 		{% block head_script %}{% endblock %}
 	</head>
 	<body>
